Fixed default shapefile CRS issue; added Reader import
Fixed an issue with the the ConvectiveOutlookFeature where it would not
plot correctly. This was due to the CRS being set incorrectly during
instantiation. The correct CRS is now implemented by default. Will have
to test if this can be overridden or dynamically read.

Added cartopy.io.shapereader.Reader import. This was needed to ensure
proper plotting.