Added hatching for SIG and Day 2
As thing currently stand, this will work and plot outlooks. The issue
is that it will not do what you expect when it does not correctly
interpolate where the bounds of the outlook should be. Not all lines
are closed and then the plotting methods will fill in the gaps,
sometimes very poorly. This version will hang around until an update
to use the recenly implemented geoJSON outlook files can be built.