Start process of using geoJSON
Mainly updating the shapereader module to grab the geoJSON instead
of the shapefiles. Changes will then need to be made to the feature
classes so that things get plotted correctly. With color information
now stord in the geoJSON attributes, helper functions are less
likely to be needed.

Secondary to this, the plan will be to merge this geoJSON version
into master and create another branch for the shapefile version
should further development occur at some point in the future.